requirements.txt
anyio==3.6.1
certifi==2022.6.15
cfgv==3.3.1
charset-normalizer==2.1.0
distlib==0.3.4
fastapi==0.78.0
filelock==3.7.1
identify==2.5.1
idna==3.3
nodeenv==1.7.0
platformdirs==2.5.2
pre-commit==2.19.0
pydantic==1.9.1
pymongo==4.1.1
python-dotenv==0.20.0
PyYAML==6.0
requests==2.28.1
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
toml==0.10.2
typing-extensions==4.3.0
urllib3==1.26.10
virtualenv==20.15.1
